Address a bunch of Micha comments
